+++ pkg/cmd/korectl/allocations.go
@@ -17,4 +17,4 @@
  * along with hub-apiserver.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package hubctl
+package korectl
