+++ pkg/cmd/korectl/commands.go
@@ -17,7 +17,7 @@
  * along with hub-apiserver.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package hubctl
+package korectl
 
 import (
 	"github.com/urfave/cli"
