+++ pkg/cmd/korectl/requestor.go
@@ -17,7 +17,7 @@
  * along with hub-apiserver.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package hubctl
+package korectl
 
 import (
 	"bytes"
